requirements.txt
PyYAML==5.4.1
jsonschema==4.18.6